+++ test/cli.js
@@ -1,5 +1,6 @@
 var bag = require('bagofholding'),
   buster = require('buster'),
+  BuildLight = require('../lib/buildlight'),
   cli = require('../lib/cli'),
   irc = require('../lib/irc'),
   Jenkins = new require('../lib/jenkins'),
@@ -21,6 +22,7 @@ buster.testCase('cli - exec', {
       assert.defined(actions.commands.irc.action);
       assert.defined(actions.commands.feed.action);
       assert.defined(actions.commands.ninja.action);
+      assert.defined(actions.commands.buildlight.action);
       done();
     };
     this.stub(bag, 'cli', { command: mockCommand });
@@ -604,6 +606,30 @@ buster.testCase('cli - ninja', {
       cb(null, 'OK');
     });
     this.stub(NinjaBlocks.prototype, 'notify', function (result) {
+      assert.equals(result, 'OK');
+      done();
+    });
+    cli.exec();
+  }
+});
+
+buster.testCase('cli - buildlight', {
+  setUp: function () {
+    this.mockConsole = this.mock(console);
+  },
+  'should notify buildlight when there is no monitoring error': function (done) {
+    this.stub(bag, 'cli', {
+      command: function (base, actions) {
+        actions.commands.buildlight.action('somejob', '* * * * * *', { scheme: 'red,green,blue', usbled: '/some/usbled/path' });
+      },
+      exitCb: bag.cli.exitCb
+    });
+    this.stub(Jenkins.prototype, 'monitor', function (jobName, schedule, cb) {
+      assert.equals(jobName, 'somejob');
+      assert.equals(schedule, '* * * * * *');
+      cb(null, 'OK');
+    });
+    this.stub(BuildLight.prototype, 'notify', function (result) {
       assert.equals(result, 'OK');
       done();
     });