+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 0.1.4-pre
-*
+* Add buildlight command
 
 ### 0.1.3
 * Fix build trigger error on Jenkins v1.5xx when job requires auth